% Auto-generated: do not edit by hand
\name{molecular}

\alias{molecular}

\title{molecular component}

\description{

}

\usage{
molecular(mol=NULL, nop_selection=NULL, onMol2DInstanceCreated=NULL,
onSelectionChanged=NULL, selection=NULL, smiles=NULL)
}

\arguments{
\item{mol}{Character. }

\item{nop_selection}{Logical. }

\item{onMol2DInstanceCreated}{}

\item{onSelectionChanged}{}

\item{selection}{Lists containing elements .
those elements have the following types:
. }

\item{smiles}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

